+++ packages/feedback/index.ts
@@ -4,3 +4,4 @@ export { Loading } from './src/Loading';
 export { Error } from './src/Error';
 export { NotFound } from './src/NotFound';
 export { GlobalError } from './src/GlobalError';
+export { NotPermitted } from './src/NotPermitted';