+++ packages/providers/index.ts
@@ -13,4 +13,5 @@ export * from './src/DispatchProvider';
 export * from './src/UserProvider';
 export * from './src/DictionaryProvider';
 export * from './src/NotificationProvider';
+export * from './src/PermissionProvider';
 // cache bust